==> 0 - Prepare the data.ipynb <==
# Data Pre-Prcoessing

## 0 - Import 

import pandas as pd
import json
from os import listdir

## 1 - Dataset

### 1.1 - Train

coleridge_df = pd.read_csv('./data/raw/train.csv')
coleridge_df

body = [] 
for ID in coleridge_df['Id'].to_numpy() : 
    with open(f'./data/raw/train/{ID}.json') as f :
        data = json.load(f)
        paper = ''
        for section in data :
            
            section_title = section['section_title']
            section_text = section['text']
            
            paper += section_text.replace('\n', ' ') + ' '
            
        body.append(paper)
        
assert len(body) == coleridge_df.__len__()
coleridge_df['body'] = body

### 1.2 - Test

test_body = [] 
test_path = './data/raw/test/'
for file in listdir(test_path) : 
    with open(test_path+file) as f :
        data = json.load(f)
        paper = ''
        for section in data :
            
            section_title = section['section_title']
            section_text = section['text']
            
            paper += section_text.replace('\n', ' ') + ' '
            
        test_body.append(paper)
        
test_dict = {
    'body' : test_body
}

coleridge_test = pd.DataFrame(test_dict)

## 4 - Drop datasets

coleridge_df.to_csv('./data/dataset/coleridge_df.csv')
coleridge_test.to_csv('./data/dataset/coleridge_test.csv')
