==> 1 - Data Pre-Processing.ipynb <==
import math
import re 
import string
from tqdm.notebook import tqdm

import pandas as pd
import numpy as np
import pickle as pkl

from sklearn.model_selection import train_test_split
from transformers import RobertaTokenizer, RobertaTokenizerFast, RobertaModel
import torch

TODO : 

    - Check Acronyms 
    - Avoid Overfitting on Dataset names

## 1 - Chunking Documents

coleridge_df = pd.read_csv('./data/dataset/coleridge_df.csv', index_col = 0)
coleridge_test = pd.read_csv('./data/dataset/coleridge_test.csv', index_col = 0)
coleridge_df = coleridge_df.rename({'body':'document', 'dataset_label' : 'answer'})
coleridge_test = coleridge_test.rename({'body':'document'})

coleridge_df

### Train - Dev

def replace_misspells_answer(document, answer) :
    
    # Handle special case by hand
    if 'Alzheimer' in document :
        document = document.replace('\'s', 's')
        
    if 'NOAA' in document : 
        document = document.replace('NOAA / National', 'NOAA National')
        document = document.replace('water-level', 'water level')
        document = document.replace('NOAA / National Water Level Observation Network',  'NOAA National Water Level Observation Network')
        
    if 'Baltimore Longitudinal Study of Aging' in document : 
        document = document.replace('Baltimore Longitudinal Study of Aging, BLSA', 'Baltimore Longitudinal Study of Aging (BLSA)')
        document = document.replace('Baltimore Longitudinal Study of Aging (BLSA; represented below the gene)', 'Baltimore Longitudinal Study of Aging (BLSA)')
        document = document.replace('Baltimore Longitudinal Study of Aging (BLSA-3T)', 'Baltimore Longitudinal Study of Aging (BLSA)')
        document = document.replace('BALTIMORE LONGITUDINAL STUDY OF AGING', 'Baltimore Longitudinal Study of Aging (BLSA)')
    
    reg = answer.replace('(', '\(', )
    reg = reg.replace(')', '\)', )
    
    # Clean punctuation surrounding dataset name
    re1 = '.' + reg + '.'
    document = re.sub(re1, ' '+answer+' ', document)
    
    # Case sensitive
    document = re.sub(reg, answer, document, flags = re.I)
    
    # Accronym punctued cleaning
    accr = answer.split()[0]
    document = re.sub('.'+accr+'.', ' '+accr+' ', document)
    
    # Replace double whitespaces
    document = document.replace('  ', ' ')
    
    return document

def clean_text(text):
    punct = string.punctuation.replace("'","")
    text = re.sub('[%s]' % re.escape(punct), ' ', str(text).lower())
    text = re.sub('[\d]', ' ', text)
    text = re.sub(' [\w]{1} ', ' ', text)
    text = re.sub('\s+([a-zA-Z]\s+)*', ' ', text)
    text = " ".join(text.split())
    return text

def find_answer(document, answer) :
    document = document.split()
    answer = answer.split()
    
    all_positions = []
    for i in range(len(document)-len(answer)+1) : 
        flag = False
        if document[i:i+len(answer)] == answer : 
            flag = True
            
        if flag == True : 
            all_positions.append([i,i+len(answer)])
                
    return all_positions

def chunking_documents(documents, answers, tokenizer, max_seq_length, sliding_window) : 
    
    documents_chunked = []
    aug_answers = []
    all_answers_pos = []
    for document, answer in tqdm(zip(documents, answers)) : 
        
        document = replace_misspells_answer(document, answer)
        
        input_ids = tokenizer.encode(document, add_special_tokens = False)
        document_length = len(input_ids)
        
        chunks, chunks_answers_pos, chunk_ans = [], [], []
        for i in range(0, math.ceil((document_length - max_seq_length) / sliding_window) + 1): 
            chunk = input_ids[i*sliding_window:(i*sliding_window) + max_seq_length]
            
            chunk_doc = tokenizer.decode(chunk)
            answer_pos = find_answer(clean_text(chunk_doc), clean_text(answer))
            
            if answer_pos == [] : 
                chunk_ans.append('')
                
            else : 
                chunk_ans.append(answer)
            
            chunks.append(chunk_doc)
            chunks_answers_pos.append(answer_pos)
            
        all_answers_pos.append(chunks_answers_pos)
        documents_chunked.append(chunks)
        n_chunks = len(chunks)
        aug_answer = [answer] * n_chunks
        aug_answers.append(aug_answer)
        
    return documents_chunked, aug_answers, all_answers_pos

documents = coleridge_df.body.to_numpy()
answers = coleridge_df.dataset_label.to_numpy()

max_seq_length = 490
overlapping_size = 50
sliding_window = max_seq_length - overlapping_size
tokenizer = RobertaTokenizerFast.from_pretrained('roberta-base')

documents_chunked, answers, answers_pos = chunking_documents(documents, answers, tokenizer, max_seq_length, sliding_window)

chunked_dataset = {
    'document' : documents_chunked,
    'answer' : answers,
    'answer_pos' : answers_pos
}

chk_coleridge_df = pd.DataFrame(chunked_dataset)

chk_coleridge_df = chk_coleridge_df.explode(['document', 'answer', 'answer_pos'])
chk_coleridge_df

def check_has_answer(answer_pos) : 
    if len(answer_pos) == 0 :
        return False 
    else : 
        return True
    
chk_coleridge_df['has_answer'] = chk_coleridge_df['answer_pos'].apply(check_has_answer)

pkl.dump(chk_coleridge_df, open('./data/dataset/chunked_coleridge_df.pkl', 'wb'))

### Test

def chunking_documents_inference(documents, tokenizer, max_seq_length, sliding_window) : 
    
    documents_chunked = []
    for document in tqdm(documents) : 
        
        input_ids = tokenizer.encode(document, add_special_tokens = False)
        document_length = len(input_ids)
        
        chunks = []
        for i in range(0, math.ceil((document_length - max_seq_length) / sliding_window) + 1): 
            chunk = input_ids[i*sliding_window:(i*sliding_window) + max_seq_length]
            
            chunk_doc = tokenizer.decode(chunk)
            
            chunks.append(chunk_doc)
            
        documents_chunked.append(chunks)
        
    return documents_chunked

documents = coleridge_test.body.to_numpy()

max_seq_length = 490
overlapping_size = 50
sliding_window = max_seq_length - overlapping_size
tokenizer = RobertaTokenizerFast.from_pretrained('roberta-base')

test_documents_chunked = chunking_documents_inference(documents, tokenizer, max_seq_length, sliding_window)

test_chunked_dataset = {
    'document' : test_documents_chunked,
}

test_chk_coleridge_df = pd.DataFrame(test_chunked_dataset)
test_chk_coleridge_df = test_chk_coleridge_df.explode(['document'])
test_chk_coleridge_df

pkl.dump(test_chk_coleridge_df, open('./data/dataset/test.pkl', 'wb'))

## 2 - Splitting Train & dev set

chk_coleridge_df = pkl.load(open('./data/dataset/chunked_coleridge_df.pkl', 'rb'))

answer = chk_coleridge_df['answer'].unique()
has_answer = chk_coleridge_df['has_answer'].to_numpy()

train_answer, dev_answer, _, _ = train_test_split(answer, answer, test_size = 0.2)
train_answer, test_answer, _, _ = train_test_split(train_answer, train_answer, test_size = 0.2)

train = chk_coleridge_df[chk_coleridge_df.answer.isin(train_answer)]
print(f'Number of texts that contain an answer : {len(train[train.has_answer == True])}')
print(f'Number of texts that do not contain an answer : {len(train[train.has_answer == False])}')

dev = chk_coleridge_df[chk_coleridge_df.answer.isin(dev_answer)]
print(f'Number of texts that contain an answer : {len(dev[dev.has_answer == True])}')
print(f'Number of texts that do not contain an answer : {len(dev[dev.has_answer == False])}')

test = chk_coleridge_df[chk_coleridge_df.answer.isin(test_answer)]
print(f'Number of texts that contain an answer : {len(test[test.has_answer == True])}')
print(f'Number of texts that do not contain an answer : {len(test[test.has_answer == False])}')

pkl.dump(train, open('./data/dataset/train.pkl', 'wb'))
pkl.dump(dev, open('./data/dataset/dev.pkl', 'wb'))
pkl.dump(test, open('./data/dataset/test.pkl', 'wb'))

## 3 - Downsampling FN

train = pkl.load(open('./data/dataset/train.pkl', 'rb'))
dev = pkl.load(open('./data/dataset/dev.pkl', 'rb'))
test = pkl.load(open('./data/dataset/test.pkl', 'rb'))

def downsampling_FN(dataframe) : 
    
    reset_df = dataframe.reset_index()
    FN = reset_df[reset_df.has_answer == False]
    TP = reset_df[reset_df.has_answer == True]
    
    n_true = len(TP)
    n_false = len(FN)
    
    keep_ratio = 1- (n_true/n_false)
    
    keep = [False if np.random.random_sample() < keep_ratio else True for k in range(n_false)]
    
    keep_index = FN.index.to_numpy()[keep]
    keep_index = np.concatenate((keep_index,TP.index.to_numpy()))
    
    downsample_df = reset_df[reset_df.index.isin(keep_index)]
    downsample_df.index = downsample_df['index']
    downsample_df = downsample_df.drop(columns = ['index'])
    
    return downsample_df
    

train_ds = downsampling_FN(train)
dev_ds = downsampling_FN(dev)
test_ds = downsampling_FN(test)

pkl.dump(train_ds, open('./data/dataset/train_ds.pkl', 'wb'))
pkl.dump(dev_ds, open('./data/dataset/dev_ds.pkl', 'wb'))
pkl.dump(test_ds, open('./data/dataset/test_ds.pkl', 'wb'))
